+++ src/Pages/Home/index.js
@@ -1,7 +1,5 @@
 import React, { useState, useEffect } from "react";
-import { Platform } from "react-native";
-import * as Notifications from "expo-notifications";
-import * as Device from 'expo-device';
+import * as Notifications from 'expo-notifications';
 import {
   AntDesign,
   FontAwesome,
@@ -27,12 +25,21 @@ import Progress from "../../Components/Progress";
 import HomeHeader from "../../Components/HomeHeader";
 import AddEarningModal from "../../Components/AddEarningModal";
 import { colors, screenHeight, screenWidth } from "../../Constants/constants";
+import registerForPushNotificationsAsync from "../../../registerForPushNotificationsAsync";
 
 const getRemaining = (time) => {
   const mins = Math.floor(time / 60);
   const secs = time - mins * 60;
   return { mins, secs };
 };
+
+Notifications.setNotificationHandler({
+  handleNotification: async () => ({
+    shouldShowAlert: true,
+    shouldPlaySound: false,
+    shouldSetBadge: false,
+  }),
+});
 
 export default function Home({ navigation }) {
   const [remainingSecs, setRemainingSecs] = useState(0);
@@ -43,34 +50,24 @@ export default function Home({ navigation }) {
     setIsActive(!isActive);
   };
 
-  registerForPushNotificationsAsync = async () => {
-    if (Device.isDevice) {
-      const { status: existingStatus } =
-        await Notifications.getPermissionsAsync();
-      let finalStatus = existingStatus;
-      if (existingStatus !== "granted") {
-        const { status } = await Notifications.requestPermissionsAsync();
-        finalStatus = status;
-      }
-      if (finalStatus !== "granted") {
-        alert("Failed to get push token for push notification!");
-        return;
-      }
-      const token = (await Notifications.getExpoPushTokenAsync()).data;
-      console.log(token);
-      this.setState({ expoPushToken: token });
-    } else {
-      alert("Must use physical device for Push Notifications");
-    }
+  state = {
+    notification: {},
+  };
 
-    if (Platform.OS === "android") {
-      Notifications.setNotificationChannelAsync("default", {
-        name: "default",
-        importance: Notifications.AndroidImportance.MAX,
-        vibrationPattern: [0, 250, 250, 250],
-        lightColor: "#FF231F7C",
-      });
-    }
+  componentDidMount = () => {
+    registerForPushNotificationsAsync();
+
+    Notifications.addNotificationReceivedListener(this._handleNotification);
+    
+    Notifications.addNotificationResponseReceivedListener(this._handleNotificationResponse);
+  };
+
+  _handleNotification = notification => {
+    this.setState({ notification: notification });
+  };
+
+  _handleNotificationResponse = response => {
+    console.log(response);
   };
 
   useEffect(() => {
